fix tests for windows int width

+++ lenskit/tests/test_pipeline_types.py
@@ -81,7 +81,7 @@ def test_data_compat_generic():
 
 
 def test_numpy_typecheck():
-    assert is_compatible_data(np.arange(10), NDArray[np.int64])
+    assert is_compatible_data(np.arange(10, dtype="i8"), NDArray[np.int64])
     assert is_compatible_data(np.arange(10, dtype="i4"), NDArray[np.int32])
     assert is_compatible_data(np.arange(10), ArrayLike)
     assert is_compatible_data(np.arange(10), NDArray[np.integer])
